--- autogpt/agent/__init__.py
@@ -1,3 +1,0 @@
-from autogpt.agent.agent import Agent
-
-__all__ = ["Agent"]
