+++ autogpt/agents/__init__.py
@@ -1,0 +1,3 @@
+from .agent import Agent
+
+__all__ = ["Agent"]
