+++ autogpt/commands/file_operations.py
@@ -8,7 +8,7 @@ import os.path
 from pathlib import Path
 from typing import Generator, Literal
 
-from autogpt.agent.agent import Agent
+from autogpt.agents.agent import Agent
 from autogpt.command_decorator import command
 from autogpt.logs import logger
 from autogpt.memory.vector import MemoryItem, VectorMemory
