+++ autogpt/commands/git_operations.py
@@ -2,7 +2,7 @@
 
 from git.repo import Repo
 
-from autogpt.agent.agent import Agent
+from autogpt.agents.agent import Agent
 from autogpt.command_decorator import command
 from autogpt.url_utils.validators import validate_url
