+++ autogpt/commands/image_gen.py
@@ -9,7 +9,7 @@ import openai
 import requests
 from PIL import Image
 
-from autogpt.agent.agent import Agent
+from autogpt.agents.agent import Agent
 from autogpt.command_decorator import command
 from autogpt.logs import logger
